+++ modding/mod_utils.h
@@ -49,6 +49,7 @@ bool EnumeratePolys(__int16 *ptrObj, ENUM_POLYS_CB callback, POLYFILTER *filter,
 
 #ifdef FEATURE_MOD_CONFIG
 bool IsModConfigLoaded();
+const char *GetModLoadingPix();
 
 void UnloadModConfiguration();
 bool LoadModConfiguration(LPCTSTR levelFilePath);